+++ src/js_schema.hpp
@@ -403,7 +403,7 @@ typename T::Object Schema<T>::object_for_property(ContextType ctx, const Propert
         }
     }
     else {
-        Object::set_property(ctx, object, type_string, Value::from_string(ctx, string_for_property_type(property.type)));
+        Object::set_property(ctx, object, type_string, Value::from_string(ctx, js_string_for_property_type(property.type)));
     }
 
     static const String object_type_string = "objectType";
@@ -411,7 +411,7 @@ typename T::Object Schema<T>::object_for_property(ContextType ctx, const Propert
         Object::set_property(ctx, object, object_type_string, Value::from_string(ctx, property.object_type));
     }
     else if (is_array(property.type)) {
-        Object::set_property(ctx, object, object_type_string, Value::from_string(ctx, string_for_property_type(property.type & ~realm::PropertyType::Flags)));
+        Object::set_property(ctx, object, object_type_string, Value::from_string(ctx, js_string_for_property_type(property.type & ~realm::PropertyType::Flags)));
     }
 
     static const String property_string = "property";